--- src/main/java/com/whales/api/dto/request/GoogleLoginRequest.java
@@ -1,8 +1,0 @@
-package com.whales.api.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record GoogleLoginRequest (
-        @NotBlank String code,
-        @NotBlank String redirectUri
-) {}
