--- src/main/java/com/whales/api/dto/request/PostRequest.java
@@ -1,19 +1,0 @@
-package com.whales.api.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PostRequest {
-
-    public UUID userId;
-    public String title;
-    public String content;
-}
