--- src/main/java/com/whales/api/dto/response/MeResponse.java
@@ -1,12 +1,0 @@
-package com.whales.api.dto.response;
-
-import java.util.UUID;
-
-public record MeResponse(
-        UUID id,
-        String email,
-        String displayName,
-        String nicknameColor,
-        String status,
-        String avatarUrl
-) {}
