--- src/main/java/com/whales/api/dto/response/TokenResponse.java
@@ -1,7 +1,0 @@
-package com.whales.api.dto.response;
-
-public record TokenResponse(
-        String accessToken,
-        long expiresIn
-) {
-}
