+++ src/main/java/com/whales/auth/api/GoogleLoginRequest.java
@@ -1,0 +1,8 @@
+package com.whales.auth.api;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record GoogleLoginRequest (
+        @NotBlank String code,
+        @NotBlank String redirectUri
+) {}
