+++ src/main/java/com/whales/auth/api/TokenResponse.java
@@ -1,0 +1,7 @@
+package com.whales.auth.api;
+
+public record TokenResponse(
+        String accessToken,
+        long expiresIn
+) {
+}
