--- src/main/java/com/whales/domain/post/ContentStatus.java
@@ -1,6 +1,0 @@
-package com.whales.domain.post;
-
-public enum ContentStatus {
-    ACTIVE,
-    INACTIVE
-}
