--- src/main/java/com/whales/domain/post/Post.java
@@ -1,59 +1,0 @@
-package com.whales.domain.post;
-
-import com.whales.domain.user.User;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.Instant;
-import java.util.UUID;
-
-@Entity
-@Table(name = "posts")
-@Getter
-@Setter
-public class Post {
-
-    @Id
-    @GeneratedValue
-    private UUID id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "author_id", nullable = false,
-            foreignKey = @ForeignKey(ConstraintMode.CONSTRAINT))
-    private User author;
-
-    @Column(length = 120, nullable = false)
-    private String title;
-
-    @Column(name = "body", columnDefinition = "TEXT", nullable = false)
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private ContentStatus status = ContentStatus.ACTIVE;
-
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private Instant createdAt = Instant.now();
-
-    @Column(name = "updated_at", nullable = false)
-    private Instant updatedAt = Instant.now();
-
-    @Column(name = "deleted_at")
-    private Instant deletedAt;
-
-    @PrePersist
-    void onCreate() {
-        this.createdAt = Instant.now();
-        this.updatedAt = Instant.now();
-    }
-
-    @PreUpdate
-    void onUpdate() {
-        this.updatedAt = Instant.now();
-    }
-
-    public boolean isDeleted() {
-        return deletedAt != null;
-    }
-}
