--- src/main/java/com/whales/domain/session/SessionRepository.java
@@ -1,8 +1,0 @@
-package com.whales.domain.session;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface SessionRepository extends JpaRepository<Session, UUID> {
-}
