--- src/main/java/com/whales/domain/user/UserRepository.java
@@ -1,10 +1,0 @@
-package com.whales.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-    Optional<User> findByEmail(String email);
-}
