--- src/main/java/com/whales/domain/user/UserRole.java
@@ -1,6 +1,0 @@
-package com.whales.domain.user;
-
-public enum UserRole {
-    USER,
-    ADMIN
-}
