--- src/main/java/com/whales/domain/user/UserStatus.java
@@ -1,7 +1,0 @@
-package com.whales.domain.user;
-
-public enum UserStatus {
-    ACTIVE,
-    INACTIVE,
-    BANNED
-}
