+++ src/main/java/com/whales/post/api/PostRequest.java
@@ -1,0 +1,19 @@
+package com.whales.post.api;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PostRequest {
+
+    public UUID userId;
+    public String title;
+    public String content;
+}
