+++ src/main/java/com/whales/post/domain/ContentStatus.java
@@ -1,0 +1,6 @@
+package com.whales.post.domain;
+
+public enum ContentStatus {
+    ACTIVE,
+    INACTIVE
+}
