+++ src/main/java/com/whales/session/domain/SessionRepository.java
@@ -1,0 +1,8 @@
+package com.whales.session.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface SessionRepository extends JpaRepository<Session, UUID> {
+}
