+++ src/main/java/com/whales/user/domain/UserRepository.java
@@ -1,0 +1,10 @@
+package com.whales.user.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+    Optional<User> findByEmail(String email);
+}
