+++ src/main/java/com/whales/user/domain/UserRole.java
@@ -1,0 +1,6 @@
+package com.whales.user.domain;
+
+public enum UserRole {
+    USER,
+    ADMIN
+}
