+++ src/main/java/com/whales/user/domain/UserStatus.java
@@ -1,0 +1,7 @@
+package com.whales.user.domain;
+
+public enum UserStatus {
+    ACTIVE,
+    INACTIVE,
+    BANNED
+}
